removing plugin from script dir
